SOQS
the lack of confidence, shyness, negligence and indecision. The main qualities of the two are to be consensual, helpful, friendly, warm, loving, tender and diplomatic. The life of this path usually revolves around partnerships, associations or professional relationships, which are all places where your qualities in human relations can shine.